Implement the low-memory rendering pipeline.
Includes some fixes to the various stages.

The low-memory rendering pipeline is enabled by default. This does not
change the decoding output, and has minor effects on speed (that can
likely be compensated for by better optimizing the various stages and
removing remaining non-rp code).

Two TODOs are still left before being able to fully remove the non-rp
code: DC upsampling and encoder-side roundtrip still use code that
belongs to the non-rendering-pipeline path.

```
x86 baseline: 54.481 MP/s
ARM baseline: geomean: 9.14383 MP/s [9.08262, 9.23499]

x86 after: 53.598 MP/s
ARM after: geomean: 9.25677 MP/s [9.21585, 9.31553]
```

+++ lib/jxl/render_pipeline/stage_noise.cc
@@ -141,7 +141,7 @@ class AddNoiseStage : public RenderPipelineStage {
   AddNoiseStage(const NoiseParams& noise_params,
                 const ColorCorrelationMap& cmap, size_t first_c)
       : RenderPipelineStage(RenderPipelineStage::Settings::Symmetric(
-            /*shift=*/0, /*border=*/2)),
+            /*shift=*/0, /*border=*/0)),
         noise_params_(noise_params),
         cmap_(cmap),
         first_c_(first_c) {}
@@ -239,8 +239,8 @@ class ConvolveNoiseStage : public RenderPipelineStage {
         rows[i] = GetInputRow(input_rows, c, i - 2);
       }
       float* JXL_RESTRICT row_out = GetOutputRow(output_rows, c, 0);
-      for (int64_t x = -RoundUpTo(xextra, Lanes(d));
-           x < (int64_t)(xsize + xextra); x += Lanes(d)) {
+      for (ssize_t x = -RoundUpTo(xextra, Lanes(d));
+           x < (ssize_t)(xsize + xextra); x += Lanes(d)) {
         const auto p00 = Load(d, rows[2] + x);
         auto others = Zero(d);
         for (ssize_t i = -2; i <= 2; i++) {
